moved AddProfilePictureController functions to ProfileController & removed it. (Trying to reduce extraneous files)
