+++ src/main/java/org/liftoff/recipebook/models/UserData.java
@@ -2,8 +2,10 @@ package org.liftoff.recipebook.models;
 
 import java.util.ArrayList;
 
+//is this class necessary? Could it be refactored?
 public class UserData {
 
+    //is this method necessary? Only used for searching?
     public static ArrayList<User> findUser(String usernameCheck, Iterable<User> allUsers) {
 
         ArrayList<User> currentUser = new ArrayList<>();
