+++ src/main/resources/templates/profile/add-profile-picture.html
@@ -5,7 +5,7 @@
 <nav th:replace="fragments :: page-header"></nav>
 <div>
     <form method="post">
-        <input type="text" name="profilePictureUrl" placeholder="Paste Profile Picture URL here">
+        <input type="text" name="profilePictureUrl" placeholder="Profile Picture URL here">
         <button class="btn btn-primary" type="submit">ADD/CHANGE</button>
     </form>
 </div>
